Added Assignment 3_3_1
Assignment:

Pascal's triangle

Pascal's triangle is a construction representing binomial coefficients. The triangle starts with
a row that has a single value of 1. Elements of each row are constructed by summing the
numbers above, to the left and right, and treating blank entries as 0. Here is an example of
the triangle with five rows:
      1
    1 1
   1 2 1
  1 3 3 1
1 4 6 4 1
Create unit tests to make sure that your program works as intended