Added Assignment 3_2_3
Assignment:

Exercise 1

Write a recursive function that takes a string as input and returns the string reversed.

Exercise 2

Write a recursive function that finds the Greatest Common Divisor of two numbers. Hint:  you can use the Euclidean algorithm.

Exercise 3

Write a recursive function to calculate the power of a number. Your recursive function should be able to handle any size of x to the power of n.

Optional Assignment 4

The towers of Hanoi

You have three towers and a set amount of disks  of different sizes which can slide onto any tower. The puzzle starts with all disks sorted in ascending order of size on one tower, with the smallest at the top. The objective is to move the entire stack to another tower by obeying the following rules:

Only one disk can be moved at a time.
Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack or on any empty rod.
No disk may be placed on top of a smaller disk