Added assignment_6.2.1
Assignment:

Exercise 1

Create a type for positive integers with following requirements

Default-initialises to 0
Initialising with a negative number reports an error to the user and cleans up correctly (printing is not a valid way to report errors to the user)
Can be copied (e.g. a = b assignment works, where a and b are instances of the type)
Has a member function to print the contents

Exercise 2

Create functions to add and substract our ex1. positive integers from each other.  The integers should saturate
at 0 and the maximum value (so they do not overflow, e.g. 42 - 51 = 0)

(optional) Add overloads to your add and substract functions so they work with normal integer values too.

Exercise 3

Create a class that counts how many instances of it exist.  Do not use global variables.  Create a function to print the count.