Added Assignment 6.3.2
Assignment:

Exercise 1

Returning to the earlier text file class, use encapsulation to hide the implementation details.

Split the implementation and declarations into source and header file(s).  Pay attention to the
headers you need to include in each.

Exercise 2

A friend function is a function that has access to the internals (private members)
of a type.  You can declare such inside a class using friend keyword.

Make the functions to add and substract work with the encapsulated version of
the positive integer class.

Exercise 3

Do the same for the Object class (6.1.2) we created earlier.