Added Assignment 6.4.1
Assignment:

Exercise 1

Create a namespace and a type in the global namespace.  Create a function in the
namespace and a static member function for the type.  Try using them from the global
namespace.

Exercise 2

Create a 3D vector type (e.g. a type with x, y and z values) in your own namespace.
Then overload the << operator so you can use that type directly with std::cout

Think about in what namespace the overload should be located (try it out, it might be a bit surprising).

Exercise 3

Overload +, -, += and -= operators for your vector type to handle vector addition
and substraction, pay attention to const-ness of your variables/parameters.