Added Assignment 7.2.1
Assignment:

Exercise 1a.

Create a function with an integer template parameters, that prints out the first integer
value, and then calls the function with remaining values.  Break the recursion using
overload resolution.

Exercise 1b.

Create a function with an integer template parameters, that prints out the first integer
value, and then calls the function with remaining values.  Break the recursion using
if contexpr

Exercise 2.

Create a function that takes arbitrary number of arguments and returns the sum of the
arguments.  How should you handle errors (such as argument not being a number) here?

Exercise 3.

In exercise 7.1.1. we created a function to check if two types are the same.  Expand on this
to check if any number of types given are all the same.

Exercise 4.

Create a function that takes arbitrary number of arguments and return the product of the
arguments times the number of arguments.

Exercise 5.

Create a function that takes arbitrary number of containers as arguments and creates a
concatenated container.  (e.g. concatenate(std::vector{1, 2}, std::list{5, 8}) -> std::vector{1, 2, 5, 8};